requirements.txt
aiohttp==3.10.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.4.0
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gspread==6.1.2
httpx==0.27.0
idna==3.7
multidict==6.0.5
oauthlib==3.2.2
openai==1.40.0
proto-plus==1.24.0
protobuf==5.27.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
